# openapi/agh.json
{
...
    "info": {
        "title": "AGH API",
        "version": "1.0.0"
    },
    "openapi": "3.0.3",
    "paths": {
...
        "/api/workspaces": {
            "get": {
                "operationId": "listWorkspaces",
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "workspaces": {
                                            "items": {
                                                "properties": {
                                                    "add_dirs": {
                                                        "items": {
                                                            "type": "string"
                                                        },
                                                        "type": "array"
                                                    },
                                                    "created_at": {
                                                        "format": "date-time",
                                                        "type": "string"
                                                    },
                                                    "default_agent": {
                                                        "type": "string"
                                                    },
                                                    "id": {
                                                        "type": "string"
                                                    },
                                                    "name": {
                                                        "type": "string"
                                                    },
                                                    "root_dir": {
                                                        "type": "string"
                                                    },
                                                    "updated_at": {
                                                        "format": "date-time",
                                                        "type": "string"
                                                    }
                                                },
                                                "required": [
                                                    "add_dirs",
                                                    "created_at",
                                                    "id",
                                                    "name",
                                                    "root_dir",
                                                    "updated_at"
                                                ],
                                                "type": "object"
                                            },
                                            "type": "array"
                                        }
                                    },
                                    "required": ["workspaces"],
                                    "type": "object"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "required": ["error"],
                                    "type": "object"
                                }
                            }
                        },
                        "description": "Internal server error"
                    },
                    "default": {
                        "description": ""
                    }
                },
                "summary": "List registered workspaces",
                "tags": ["workspaces"],
                "x-agh-transports": ["http", "uds"]
            },
            "post": {
                "operationId": "createWorkspace",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "add_dirs": {
                                        "items": {
                                            "type": "string"
                                        },
                                        "type": "array"
                                    },
                                    "default_agent": {
                                        "type": "string"
                                    },
                                    "name": {
                                        "type": "string"
                                    },
                                    "root_dir": {
                                        "type": "string"
                                    }
                                },
                                "required": ["root_dir"],
                                "type": "object"
                            }
                        }
                    },
                    "description": "JSON request body",
                    "required": true
                },
                "responses": {
                    "201": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "workspace": {
                                            "properties": {
                                                "add_dirs": {
                                                    "items": {
                                                        "type": "string"
                                                    },
                                                    "type": "array"
                                                },
                                                "created_at": {
                                                    "format": "date-time",
                                                    "type": "string"
                                                },
                                                "default_agent": {
                                                    "type": "string"
                                                },
                                                "id": {
                                                    "type": "string"
                                                },
                                                "name": {
                                                    "type": "string"
                                                },
                                                "root_dir": {
                                                    "type": "string"
                                                },
                                                "updated_at": {
                                                    "format": "date-time",
                                                    "type": "string"
                                                }
                                            },
                                            "required": [
                                                "add_dirs",
                                                "created_at",
                                                "id",
                                                "name",
                                                "root_dir",
                                                "updated_at"
                                            ],
                                            "type": "object"
                                        }
                                    },
                                    "required": ["workspace"],
                                    "type": "object"
                                }
                            }
                        },
                        "description": "Created"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "required": ["error"],
                                    "type": "object"
                                }
                            }
                        },
                        "description": "Invalid workspace request"
                    },
                    "409": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "required": ["error"],
                                    "type": "object"
                                }
                            }
                        },
                        "description": "Workspace conflict"
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "required": ["error"],
                                    "type": "object"
                                }
                            }
                        },
                        "description": "Internal server error"
                    },
                    "default": {
                        "description": ""
                    }
                },
                "summary": "Register a workspace",
                "tags": ["workspaces"],
                "x-agh-transports": ["http", "uds"]
            }
        },
        "/api/workspaces/resolve": {
            "post": {
                "operationId": "resolveWorkspace",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "path": {
                                        "type": "string"
                                    }
                                },
                                "required": ["path"],
                                "type": "object"
                            }
                        }
                    },
                    "description": "JSON request body",
                    "required": true
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "workspace": {
                                            "properties": {
                                                "add_dirs": {
                                                    "items": {
                                                        "type": "string"
                                                    },
                                                    "type": "array"
                                                },
                                                "created_at": {
                                                    "format": "date-time",
                                                    "type": "string"
                                                },
                                                "default_agent": {
                                                    "type": "string"
                                                },
                                                "id": {
                                                    "type": "string"
                                                },
                                                "name": {
                                                    "type": "string"
                                                },
                                                "root_dir": {
                                                    "type": "string"
                                                },
                                                "updated_at": {
                                                    "format": "date-time",
                                                    "type": "string"
                                                }
                                            },
                                            "required": [
                                                "add_dirs",
                                                "created_at",
                                                "id",
                                                "name",
                                                "root_dir",
                                                "updated_at"
                                            ],
                                            "type": "object"
                                        }
                                    },
                                    "required": ["workspace"],
                                    "type": "object"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "required": ["error"],
                                    "type": "object"
                                }
                            }
                        },
                        "description": "Invalid workspace path"
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "required": ["error"],
                                    "type": "object"
                                }
                            }
                        },
                        "description": "Workspace not found"
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "required": ["error"],
                                    "type": "object"
                                }
                            }
                        },
                        "description": "Internal server error"
                    },
                    "default": {
                        "description": ""
                    }
                },
                "summary": "Resolve or register a workspace from a path",
                "tags": ["workspaces"],
                "x-agh-transports": ["http", "uds"]
            }
        },
...
    },
...
}